+++ desktop/cores/sdl/CycleCounter.h
@@ -1,0 +1,18 @@
+#ifndef _CycleCounter_h_
+#define _CycleCounter_h_
+
+void klst_enable_cycle_counter() {
+}
+
+uint32_t klst_get_cycles() {
+    return 0;
+}
+
+void klst_reset_cycles() {
+}
+
+float klst_cyclesToMicros(uint32_t pCycles) {
+  return 0;
+}
+
+#endif // _CycleCounter_h_